+++ lib/src/screen.dart
@@ -7,57 +7,61 @@ part of '../griddle.dart';
 /// - Synchronize the state of the (internal) screen with an external surface.
 ///
 /// There is only one provided way to create a screen:
-/// - [Screen.terminal], which creates a screen that interfaces with a terminal.
+/// - [Screen.output], which creates a screen that uses a low-level [RawScreen].
 ///
 /// However, [Screen] exists as an abstraction: _extend_ and create your own!
-///
-/// **NOTE**: In the future, the screen API will also support _input_ (events).
 @sealed
-abstract class Screen extends Buffer {
-  final int _framesPerSecond;
-
-  /// Creates a disconnected in-memory screen of initial [width] and [height].
-  Screen(
-    super.width,
-    super.height, {
-    int framesPerSecond = 30,
-  }) : _framesPerSecond = RangeError.checkNotNegative(
-          framesPerSecond,
-          'framesPerSecond',
-        );
-
-  /// Creates a screen that interfaces with an external [terminal].
+abstract class Screen implements Buffer {
+  /// Creates a screen that interfaces with an external [screen].
   ///
-  /// The simplest possible "real" terminal is [Terminal.usingStdio]:
-  /// ```dart
-  /// void main() {
-  ///   final screen = Screen.terminal(Terminal.usingStdio());
-  ///
-  ///   // Use the 'screen' attached to stdin and stdout.
-  /// }
-  /// ```
-  ///
-  /// The width and height of the screen are determined by the terminal.
-  factory Screen.terminal(
-    Terminal terminal, {
-    int framesPerSecond,
-  }) = _TerminalScreen;
+  /// See [RawScreen.fromAnsiTerminal] and [RawScreen.fromStringBuffer].
+  factory Screen.output(RawScreen screen) = _Screen;
 
   /// Given the current state of the buffer, updates an external surface.
   void update();
+}
 
-  /// A stream that fires every frame the screen could be updated.
-  ///
-  /// The event value provided ([Duration]) is the time elapsed since the last
-  /// frame was emitted.
-  @nonVirtual
-  Stream<Duration> get onFrame {
-    final stopwatch = Stopwatch()..start();
-    final duration = Duration(milliseconds: 1000 ~/ _framesPerSecond);
-    return Stream.periodic(duration, (_) {
-      final time = stopwatch.elapsed;
-      stopwatch.reset();
-      return time;
-    });
+/// Simple implementation of [Screen] that delegates to a [RawScreen].
+class _Screen extends Buffer implements Screen {
+  final RawScreen _output;
+
+  _Screen(this._output) : super(_output.width, _output.height);
+
+  @override
+  void update() {
+    _output.clearScreen();
+
+    for (var i = 0; i < height; i++) {
+      _output.writeByte(0xa);
+      for (var j = 0; j < width; j++) {
+        final cell = get(j, i);
+        _writeAnsiColorSequences(cell);
+        _output.writeByte(cell.character);
+      }
+    }
+
+    _output
+      ..writeByte(0xa)
+      ..flush();
+  }
+
+  void _writeAnsiColorSequences(Cell cell) {
+    var wroteSequence = false;
+
+    final foreground = cell.foregroundColor;
+    if (foreground != null) {
+      _output.setForegroundColor(foreground);
+      wroteSequence = true;
+    }
+
+    final background = cell.backgroundColor;
+    if (background != null) {
+      _output.setBackgroundColor(background);
+      wroteSequence = true;
+    }
+
+    if (!wroteSequence) {
+      _output.resetStyles();
+    }
   }
 }